Added MIT LICENSE
example_experiment is changed slightly so it can run out of the box without a WandB API key as long as requirements.txt are met.
The experiment loop is shortened for a quick run for testing if the code can run successfully.
This distribution of source code is prepared for IEEE Code Ocean.
MIT LICENCE is applied to the opened source code.